+++ jest.setup.js
@@ -1,1 +1,3 @@
 jest.mock("react-native/Libraries/EventEmitter/NativeEventEmitter");
+
+jest.doMock("./src/common/logging");
