+++ src/common/__mocks__/logging.ts
@@ -1,0 +1,3 @@
+export const log = jest.fn();
+export const logError = jest.fn();
+export const universalCatch = jest.fn();
