+++ src/common/logging.ts
@@ -1,7 +1,7 @@
 import crashlytics from "@react-native-firebase/crashlytics";
 import { isError } from "lodash";
 
-import { NativeFirebaseError } from "../types/firebaseTypes";
+import { isFirebaseError } from "../types/firebaseTypes";
 
 export function log(message: string | boolean | number | object, level: "trace" | "debug" | "log" | "info" | "warn" | "error" = "log") {
   try {
@@ -36,32 +36,6 @@ export function logError(error: Error) {
   } catch (error) {
     console.error(error);
   }
-}
-
-export function isFirebaseError(error: unknown): error is NativeFirebaseError {
-  if (typeof error !== "object" || error == null) {
-    return false;
-  }
-  if (typeof (error as NativeFirebaseError).code !== "string") {
-    return false;
-  }
-  if (typeof (error as NativeFirebaseError).message !== "string") {
-    return false;
-  }
-  if (typeof (error as NativeFirebaseError).name !== "string") {
-    return false;
-  }
-  if (typeof (error as NativeFirebaseError).namespace !== "string") {
-    return false;
-  }
-  if (typeof (error as NativeFirebaseError).stack !== "string" && (error as NativeFirebaseError).stack != null) {
-    return false;
-  }
-  if (typeof (error as NativeFirebaseError).cause !== "string" && (error as NativeFirebaseError).cause != null) {
-    return false;
-  }
-
-  return true;
 }
 
 export function universalCatch(error: unknown) {
