+++ src/types/firebaseTypes.ts
@@ -4,6 +4,31 @@ import { FirebaseStorageTypes } from "@react-native-firebase/storage";
 import { HourInstructionsType, SpecialComponentType } from "./hourScreenTypes";
 
 export type NativeFirebaseError = Parameters<FirebaseStorageTypes.TaskSnapshotObserver["error"]>[0];
+export function isFirebaseError(error: unknown): error is NativeFirebaseError {
+  if (typeof error !== "object" || error == null) {
+    return false;
+  }
+  if (typeof (error as NativeFirebaseError).code !== "string") {
+    return false;
+  }
+  if (typeof (error as NativeFirebaseError).message !== "string") {
+    return false;
+  }
+  if (typeof (error as NativeFirebaseError).name !== "string") {
+    return false;
+  }
+  if (typeof (error as NativeFirebaseError).namespace !== "string") {
+    return false;
+  }
+  if (typeof (error as NativeFirebaseError).stack !== "string" && (error as NativeFirebaseError).stack != null) {
+    return false;
+  }
+  if (typeof (error as NativeFirebaseError).cause !== "string" && (error as NativeFirebaseError).cause != null) {
+    return false;
+  }
+
+  return true;
+}
 
 export interface FirestoreHour {
   hourNumber: number;
